refactor: move renderer components
