+++ apps/mc-pack-tool/.eslintrc.cjs
@@ -1,0 +1,16 @@
+module.exports = {
+  env: {
+    browser: true,
+    es6: true,
+    node: true,
+  },
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:import/recommended',
+    'plugin:import/electron',
+    'plugin:import/typescript',
+  ],
+  parser: '@typescript-eslint/parser',
+};
