+++ apps/mc-pack-tool/package.json
@@ -9,7 +9,7 @@
     "package": "electron-forge package",
     "make": "electron-forge make",
     "publish": "electron-forge publish",
-    "lint": "eslint --config .eslintrc.json --ext .ts,.tsx .",
+    "lint": "eslint --ext .ts,.tsx .",
     "test": "vitest run",
     "dev": "electron-forge start"
   },
