+++ apps/mc-pack-tool/src/renderer/components/AssetBrowser.tsx
