+++ apps/mc-pack-tool/src/renderer/components/AssetSelector.tsx
